Remove 'offline' references when steps can be run in the container
The previous updates to the face detection code and Pipfile dependencies should ensure that all face detection options/steps can be run in the api container, though it may still be desirable to run the initial detection step outside the container, because it takes quite a while. So there's no need to keep around filenames and doc and code references to running most of the face detection pipeline "offline."
