+++ README.md
@@ -72,16 +72,16 @@ Steps marked with an asterisk `*` may be optional if a viable output file is pre
 1. `just add-video-4dh Video_File_Name.ext` - Creates a DB entry for the video, then locates the PHALP/4D-Humans pose estimation output file as `Video_File_Name.ext.phalp.pkl` and imports the data into the DB.
 1. **DEPRECATED**: `just add-video Video_File_Name.ext` - Creates a DB entry for the video, then locates the Open PifPaf pose estimation output file as `Video_File_Name.ext.openpifpaf.json` and imports the data into the DB.
 1. `just detect-shots Video_File_Name.ext` - Runs shot detection on each frame of the video file and writes the output to `Video_File_Name.ext.shots.TransNetV2.pkl`.\*
-1. `just add-shots Video_File_Name.ext` - Looks for a shot detection output file named `Video_File_Name.ext.shots.TransNetV2.pkl` generated for the video via the `app/detect_shots_offline.py` script in the repo and imports the data into the DB.
+1. `just add-shots Video_File_Name.ext` - Looks for a shot detection output file named `Video_File_Name.ext.shots.TransNetV2.pkl` generated for the video via the `api/detect_shots.py` script in the repo and imports the data into the DB.
 1. **DEPRECATED**: `just add-tracks Video_File_Name.ext` - Applies object tracking to the pose position data in the DB to join adjacent poses by the same figure into "tracks" that are saved in the DB. NOTE: this is only necessary when working with Open PifPaf pose data; the PHALP/4D-Humans output already includes tracking data.
 1. `just do-poem-embeddings Video_File_Name.ext` - Exports the normalized COCO 13-keypoint poses for a video into a CSV file (at `api/poem_files/Video_File_Name.ext.csv`) to be used as input when creating view-invariant pose embeddings via the Pr-VIPE model from the [POEM project](https://sites.google.com/view/pr-vipe), then runs the POEM view-invariant pose embedding generation software on the CSV, producing an output CSV file that is then loaded into the DB to create a view-invariant search index on the poses.
 1. `just add-motion Video_File_Name.ext` - Segments pose tracks into "movelets" consisting of the average position of consecutive poses that occur within 1/6 of a second of each other, then computes the degree of motion between adjacent movelets.
 1. `just load-actions Video_File_Name.ext.lart.pkl` - Loads externally generated action recognition output data for a given video from a file named `Video_File_name.ext.lart.pkl`.\*
 1. `just calculate-pose-interest Video_File_Name.ext` - Compares each pose to the global average pose for the video, computes the degree to which the poses in each frame deviate from the average, and represents the frame's "interest" level as the maximum of these values.
 1. `just calculate-action-interest Video_File_Name.ext` - If action recognition data has been loaded previously for a video, compares each pose's action vector to the global average for the video, computes the degree to which the poses in each frame deviate from the average, and represents the frame's "interest" level as the maximum of these values.\*
-1. `just detect-faces Video_File_Name.ext` - Runs face detection on each frame of the video file and writes the output to `Video_File_Name.faces.ArcFace.jsonl`. Note that this can take a very long time to run (as long as the initial offline pose estimation task) and may not be supported by the API container environment; you may prefer to run the `api/detect_faces_offline.py` script separately as a batch job.\*
-1. `just match-faces Video_File_Name.ext` - Looks for a face detection output file named `Video_File_Name.ext.faces.ArcFace.jsonl` generated for the video via the `app/detect_faces_offline.py` script in the repo, matches detected poses with detected faces and adds the information to the DB.
-1. `**OPTION 1**: just detect-labeled-faces Video_File_Name.ext` - Assumes you have placed headshot image files in `api/face_images/Video_File_Name.ext`, one for each face you want to search for in the video, named byh the associated role or actor, e.g., "Julius_Caesar.png". This runs face detection on the headshots and produces a .json summary of their features in the same folder. May also need to be run outside the API container environment, similar to `just detect-faces` above.
+1. `just detect-faces Video_File_Name.ext` - Runs face detection on each frame of the video file and writes the output to `Video_File_Name.faces.ArcFace.jsonl`. Note that this can take a very long time to run (as long as the initial offline pose estimation task). You may prefer to run the `api/detect_faces.py` script separately as a batch job.\*
+1. `just match-faces Video_File_Name.ext` - Looks for a face detection output file named `Video_File_Name.ext.faces.ArcFace.jsonl` generated for the video via the `api/detect_faces.py` script in the repo, matches detected poses with detected faces and adds the information to the DB.
+1. `**OPTION 1**: just detect-labeled-faces Video_File_Name.ext` - Assumes you have placed headshot image files in `api/face_images/Video_File_Name.ext`, one for each face you want to search for in the video, named byh the associated role or actor, e.g., "Julius_Caesar.png". This runs face detection on the headshots and produces a .json summary of their features in the same folder.
 1. `**OPTION 1**: just match-labeled-faces Video_File_Name.ext` - If face detection output has been loaded into the DB via `just match-faces` and `just detect-faces`, and `just detect-labeled-faces` also has been run, this runs a similarity analysis between detected faces in the video and the target face features corresponding to the images in `api/face_images/Video_File_Name.ext` and saves the results to the DB to create a timeline of occurrences of the target faces in the video.
 1. `**OPTION 2**: just cluster-faces Video_File_Name.ext [NUMBER_OF_PERSONS]` - If face detection output has been loaded into the DB via `just match-faces` and `just detect-faces`, performs K-means clustering on the face features and saves the results to the DB to create a timeline of every (suspected) occurrence of a person in the video. NUMBER_OF_PERSONS is approximately the total number of individual faces expected to be in the video (which may or may not correspond to the size of the cast).
 1. `just cluster-poses Video_File_Name.ext [NUMBER_OF_POSES]` - Performs K-means clustering on representative poses in tracked movelets, assuming the specified NUMBER_OF_POSES pose archetypes, and saves the results to the DB to enable an interactive timeline display of the pose cluster membership.
