+++ api/detect_shots.py
