+++ api/match_faces_to_poses.py
@@ -1,16 +1,20 @@
 #!/usr/bin/env python3
 
-"""CLI to add face detection data to OpenPifPaf pose data for a video in the db."""
+"""CLI to add face detection data to pose data for a video in the db."""
 
 import argparse
 import asyncio
 import logging
 from pathlib import Path
 
+import jsonlines
 import numpy as np
 from rich.logging import RichHandler
 
 from mime_db import MimeDb
+
+BATCH_SIZE = 1000
+FACE_FEATURES = 512  # Previously used DeepFace, which has 4096
 
 
 def unflatten_pose_data(keypoints):
@@ -42,6 +46,85 @@ def get_bbox_overlap(box1, box2):
     return 0
 
 
+async def match_faces_in_frames(video_id, faces_to_match, min_frameno, max_frameno, db):
+    logging.info(
+        f"Running match_faces_in_frames with start frame {min_frameno} end {max_frameno}"
+    )
+
+    matches_to_assign = []
+
+    frame_poses = await db.get_frame_data_range(video_id, min_frameno, max_frameno)
+
+    poses_by_frame = {}
+    for pose in frame_poses:
+        if pose["frame"] in poses_by_frame:
+            poses_by_frame[pose["frame"]].append(pose)
+        else:
+            poses_by_frame[pose["frame"]] = [pose]
+
+    for frameno in faces_to_match:
+        if frameno not in poses_by_frame:
+            continue
+        frame_faces = faces_to_match[frameno]
+        frame_poses = poses_by_frame[frameno]
+
+        for pose in frame_poses:
+            if pose["track_id"] == 0:
+                continue
+            pose_head = extract_trustworthy_coords(pose["keypoints"])[:5, :]
+            if np.isnan(pose_head).all():
+                continue
+            head_bbox = [
+                np.nanmin(pose_head[:, 0]),
+                np.nanmin(pose_head[:, 1]),
+                np.nanmax(pose_head[:, 0]) - np.nanmin(pose_head[:, 0]),
+                np.nanmax(pose_head[:, 1]) - np.nanmin(pose_head[:, 1]),
+            ]
+            if np.isnan(head_bbox).any():
+                continue
+
+            best_match = None
+            best_overlap = 0
+            for f, face in enumerate(frame_faces):
+                bbox_overlap = get_bbox_overlap(face["bbox"], head_bbox)
+                if bbox_overlap > best_overlap:
+                    best_match = f
+                    best_overlap = bbox_overlap
+
+            if best_match is not None and best_overlap > 0:
+                landmarks_vector = [
+                    coord
+                    for pair in frame_faces[best_match]["landmarks"].values()
+                    for coord in pair
+                ]
+
+                embedding = frame_faces[best_match]["embedding"]
+                # Previously, we padded all embeddings to 4096 elements because we
+                # might use DeepFace, which produces that many. But the 512
+                # elements (from ArcFace) are sufficient and use less storage.
+                if len(embedding) != FACE_FEATURES:
+                    logging.error(
+                        f"Only face embeddings with {FACE_FEATURES} dimensions are supported."
+                    )
+                    return
+
+                matches_to_assign.append(
+                    [
+                        video_id,
+                        frameno,
+                        pose["pose_idx"],
+                        frame_faces[best_match]["bbox"],
+                        frame_faces[best_match]["confidence"],
+                        landmarks_vector,
+                        embedding,
+                        pose["track_id"],
+                    ]
+                )
+
+    if len(matches_to_assign) > 0:
+        await db.add_pose_faces(matches_to_assign)
+
+
 async def main() -> None:
     """Command-line entry-point."""
 
@@ -71,8 +154,8 @@ async def main() -> None:
         handlers=[RichHandler(rich_tracebacks=True)],
     )
 
-    # This should really just be the name of the video, since its pose data
-    # should already be in the DB by the time this is run
+    faces_file = Path(f"{args.video_name}.faces.ArcFace.jsonl")
+
     video_name = Path(args.video_name)
 
     # Connect to the database
@@ -81,52 +164,53 @@ async def main() -> None:
     # Get video metadata
     video_name = video_name.name
     video_id = await db.get_video_id(video_name)
-    video_id = video_id[0]["id"]
+
+    track_frame_records = await db.get_track_frames(video_id)
+
+    track_frame_ids = {frame_record["frame"] for frame_record in track_frame_records}
+
+    faces_to_match = {}
+    min_frameno = None
+    max_frameno = None
 
     logging.info("Matching tracked poses to faces detected in video")
 
-    track_frames = await db.get_track_frames(video_id)
-
-    for track_frame in track_frames:
-        logging.info(f"matching tracked poses to faces in frame {track_frame['frame']}")
-
-        frame_faces = await db.get_frame_faces(video_id, track_frame["frame"])
-        frame_poses = await db.get_frame_data(video_id, track_frame["frame"])
-
-        for pose in frame_poses:
-            if pose["track_id"] == 0:
-                continue
-            pose_head = extract_trustworthy_coords(pose["keypoints"])[:5, :]
-            head_bbox = [
-                np.nanmin(pose_head[:, 0]),
-                np.nanmin(pose_head[:, 1]),
-                np.nanmax(pose_head[:, 0]) - np.nanmin(pose_head[:, 0]),
-                np.nanmax(pose_head[:, 1]) - np.nanmin(pose_head[:, 1]),
-            ]
-            if np.isnan(head_bbox).any():
+    with jsonlines.open(faces_file) as reader:
+        for face in reader:
+            if (
+                face["confidence"] == 0
+                or not face["landmarks"]
+                or face["frame"] not in track_frame_ids
+            ):
                 continue
 
-            best_match = None
-            best_overlap = 0
-            for f, face in enumerate(frame_faces):
-                bbox_overlap = get_bbox_overlap(face["bbox"], head_bbox)
-                if bbox_overlap > best_overlap:
-                    best_match = f
-                    best_overlap = bbox_overlap
+            if face["frame"] in faces_to_match:
+                faces_to_match[face["frame"]].append(face)
+            else:
+                faces_to_match[face["frame"]] = [face]
 
-            if best_match is not None and best_overlap > 0:
-                await db.assign_pose_face(
-                    video_id,
-                    track_frame["frame"],
-                    pose["pose_idx"],
-                    frame_faces[best_match]["bbox"],
-                    frame_faces[best_match]["confidence"],
-                    frame_faces[best_match]["landmarks"],
-                    frame_faces[best_match]["embedding"],
+            if min_frameno is None:
+                min_frameno = face["frame"]
+            else:
+                min_frameno = min(min_frameno, face["frame"])
+
+            if max_frameno is None:
+                max_frameno = face["frame"]
+            else:
+                max_frameno = max(max_frameno, face["frame"])
+
+            if len(faces_to_match) >= BATCH_SIZE:
+                await match_faces_in_frames(
+                    video_id, faces_to_match, min_frameno, max_frameno, db
                 )
-                logging.info(
-                    f"matched face {frame_faces[best_match]['bbox']} to pose {pose['pose_idx']}, track {pose['track_id']} with head bbox {head_bbox}"
-                )
+                faces_to_match = {}
+                min_frameno = None
+                max_frameno = None
+
+        if len(faces_to_match) > 0:
+            await match_faces_in_frames(
+                video_id, faces_to_match, min_frameno, max_frameno, db
+            )
 
 
 if __name__ == "__main__":
